creates a Climb route components for the show page and acesses the route id
